Modify intent to support the adaptive icon
Add the interface ArcActivityIconsToImageSkias in app_icon_factory.cc to
load a group of icons for the ARC activity.

Add ArcActivityAdaptiveIconImpl to generate the adaptive icons for the
ARC activity by calling ArcActivityIconsToImageSkias. Modify ArcApps to
set the ArcActivityAdaptiveIconImpl as the delegate for the
ActivityIconLoader if the adaptive icon is enabled.

Modify start_smart_selection_action_menu to generate the adaptive icons
by calling ArcActivityIconsToImageSkias.

BUG=1083331

Change-Id: I64a14c5459676593ceeedb442894ca9abe2cebe9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2305816
Reviewed-by: David Jacobo <djacobo@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792587}