Use ScrimCoordinator in place of ScrimView in autobot
This patch replaces the obsolete ScrimView widget with the new scrim
component in the autofill assistant. Some of the tests that included
waiting for elements to be removed from the test web page have been
updated to use criteria helpers, making them a bit less flaky.

This patch also includes a change to how the touch delegate is
attached to the scrim: it is now only attached if a delegate is set
as it would otherwise interfere with the gesture detector.

Bug: 1062099
Change-Id: I79547256efa62eef7bc1d86090942a8471a60285
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2308454
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/master@{#792680}