[profiles] Add profile colors into ProfileAttributesEntry
Upcoming updates to the multi-profile experience make the "main" color
of the profile more prominent on several surfaces. One of these surfaces
is the profile picker window which should have access to colors
attributed to unloaded profiles.

Currently, the profile theme info is stored in profile preferences which
are read from the disk when a profile is loaded. This CL caches required
colors in ProfileAttributesStorage. Data from ProfileAttributesStorage
can be accessed without having to actually load a profile from disk.

The colors are updated every time the profile's theme changes. This is
implemented by listening for the
chrome::NOTIFICATION_BROWSER_THEME_CHANGED notification in
ProfileThemeUpdateService.

Bug: 1102384
Change-Id: I4683ded5d7182117dd2a712437f64f1f00ceed72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2317796
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792683}