Create SendRegistrationPing in UpdateClient
Create a new state and event within the component type. Within the
UpdateEngine and UpdateClient create SendRegistrationPing that calls
for registration event to be sent. The execution path for the
the registration ping is not taken by the code inside the component
updater The change is needed for the code in //chrome/updater.

Bug:1096733

Change-Id: I0f239d8d837eb22dda49165ae04698d896860578
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2290195
Reviewed-by: Tommy Li <tommycli@chromium.org>
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Commit-Queue: Samantha Sappenfield <t-sasap@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#792810}