Include otool for Telemetry
Includes the hermetic Xcode binary directory as a data dependency for
Telemetry tests on Mac. Xcode is required in order to symbolize
minidumps on Mac, and this previously worked by accident due to Xcode
previously being installed system-wide. generate_breakpad_symbols.py
already supports using the hermetic binary, so we just need to ensure
that it's available on a swarming task.

Also re-enables the tests that were disabled due to the missing Xcode
binaries.

Bug: 1110441
Change-Id: Icb703e08c65c27c4b6300710033be138a0bd4da3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2324883
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792823}