Add UseCounters for Conversion Measurement API
Adds Blink Usecounters for impression registration, conversion
registration, and a composite usecounter for both of those events.
These are necessary to measure the percentage of page loads using the
API, for running an origin trial.

A new test harness is added in the chrome/ layer to verify that
metrics are recorded as expected.

Bug: 1109079
Change-Id: I2f70bb3a2eac60ff45a5654169b4db959fd1d3f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2314843
Commit-Queue: John Delaney <johnidel@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792856}