[Web Payment] Avoid sending WebContents* to service worker
PaymentAppProviderImpl was recently changed[1] to use WebContents*
instead of BrowserContext*, in preparation for making the object owned
by WebContents. However, the service worker logic still needs to use a
BrowserContext* to register callbacks in case it outlives the
WebContents, which is possible in the current state where it's a
singleton. This patch makes it so all service worker logic uses a
BrowserContext*.

Bug: 1110207
Change-Id: Iee16d9cdfc81e2f186226e11be379890d6d66eef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2324041
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Nick Burris <nburris@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792901}