[ozone/wayland] Use WaylandSurface* to identify a WaylandSurface
AcceleratedWidget used to have 1-1 relation with a wl_surface because a
platform window was 1-1 with wl_surface. This is no longer the case with
overlay forwarding where a platform window can consist of multiple
wl_surfaces.

With this CL ui::WaylandSurface* identifies a WaylandSurface,
gfx::AcceleratedWidget identifies a WaylandWindow.

This is 1/? CL for overlay forwarding using wl_subsurface.

Bug: 1063865
Change-Id: I0636817dbc5533fd8f4e5f6890bf4deac365ab2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2314138
Commit-Queue: Kramer Ge <fangzhoug@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Maksim Sisov (GMT+3) <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#792905}