Don't show app menu if anchor view isn't attached to window
Speculative fix for WindowManager$BadTokenException when trying to show
the app menu. Internally PopupWindow uses the supplied parent view to
get a window token, so check that the anchor view / it's root view are
attached to a Window in addition to checking if the decor view is
attached.

BUG=1105831

Change-Id: Ie270a6395c541a9be508f161913fb1ef79c0c74f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2327316
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Commit-Queue: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792955}