exo: Remove incorrect validation code from xdg-shell
In the unstable protocol, the anchor and gravity fields are bitfields
and we have code to check that the combination of values makes
sense. In the stable version this has been changed to an enum so
invalid combinations are no longer possible, and the code that checks
for them becomes incorrect.

Bug: 1110101
Bug: 1022716
Change-Id: I8f81a0b3f0e76a11fc0bc096724b8b3766cab399
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2326853
Commit-Queue: Fergus Dall <sidereal@google.com>
Commit-Queue: Mitsuru Oshima <oshima@chromium.org>
Auto-Submit: Fergus Dall <sidereal@google.com>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793008}