Refactor WebDialogView closing logic
WebDialogDelegate has two methods which can block the dialog from being
closed: CanCloseDialog() and OnDialogCloseRequested(). CanCloseDialog()
is a const method and the name suggests that it should simply be a
predicate without side effect. OnDialogCloseRequested(), before this cl,
is called too late in the closing process, so you will end up with a
half-broken dialog when you return false in it to block closing because
the web contents would have been closed.

This CL consolidates the two by 1) removing CanCloseDialog(); 2) Call
OnDialogCloseRequested() in WebDialogView::OnWindowCloseRequested(),
which makes OnDialogCloseRequested() actually useful. Now we are able to
move logic in Crostini{Installer,Upgrader}Dialog::CanCloseDialog(),
which requests the web page to conduct closing logic, to
OnDialogCloseRequested().

login_screen_extension_ui::DialogDelegate::CanCloseDialog() and
AddSupervisionDialog::CanCloseDialog() are also moved to the
corresponding OnDialogCloseRequested() methods.

In addition, the old AddSupervisionDialog::OnDialogCloseRequested() is
moved to DeprecatedOnDialogCloseRequested(). It seems that in the old
implementation, OnDialogCloseRequested() was not called when the user
signed out with the dialog open, and AddSupervisionDialog was relying on
this behaivor. So a new function DeprecatedOnDialogCloseRequested() is
added for now to maintain the behavior.

Bug: 970196
Change-Id: I4e552ed02d2db96886b87534ad7f187844ed04e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2278793
Commit-Queue: Jason Lin <lxj@google.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793009}