Expand occluding damage underlay optimization
This change adds a |no_damage| flag to SharedQuadState that viz
clients can use to indicate that the quad doesn't contribute damage
to the render pass. Exo is updated to use this flag. The new flag is
then used in SurfaceAggregator to expand the occluding damage
optimization to apply to any quad which is explicitly marked as the
only damaged quad in its render pass.

Bug: 1082082
Change-Id: I75b01ad724c6f0227cc1076185f301237310bc44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275283
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Maggie Chen <magchen@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793057}