Remove code to delete unused Safe Browsing lists from disk
Also removes the corresponding histograms.
This code was added to remove lists that were in use previously
but are no longer in use. The code to delete the lists was
added in 2018 and we wanted to leave it in for a few releases
to ensure that it is removed from all users' machines.

http://screen/VCRd5VUL1B7 shows those stores exist on disk for an
extremely small proportion of users.

Fixed: 1089439, 695081
Change-Id: I274707f312ec1c1c2baedf8997643ac3b95f1ba2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2327850
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Commit-Queue: Mark Pearson <mpearson@chromium.org>
Auto-Submit: Varun Khaneja <vakh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793222}