Add feedback item to power menu on login screen.
NewWindowDelegate::OpenFeedbackPage can only be used to open the
feedback dialog in the active user session.
SigninScreenHandler::HandleSendFeedback is used to open the feedback
dialog in the login screen. Due to the effort to create the instance of
SigninScreenHandler, this cl triggers the same handler associated with
the feedback accelerator from the login screen to bring up the feedback
dialog.

Bug: 1107188
Change-Id: I847f96a958f50b1ebbea6f5413e80e0002155042
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2325274
Commit-Queue: Min Chen <minch@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793225}