[AudioService] Enable sandbox based on feature flag
Prior to this CL, the method
ContentBrowserClient::ShouldSandboxAudioService() returned false,
which made the sandbox disabled by default for embedders that
do not override this method, like content_shell.
Instead, this should return true if the AudioServiceSandbox feature
is enabled.

This was a regression introduced in http://crrev.com/790723

Bug: 1111349,1109346
Change-Id: Ib38c02e63ee039e6550707026cf15a1ce8739575
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2329891
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793233}