Make SquaredDistanceBetweenRects() distinguish intersection.
When DIPToScreenRect() is asked to find the nearest display to a given rect, it uses the minimum SquaredDistanceBetweenRects() to choose.  Since adjacent and overlapping rectangles had the same reported distance (0), a rect on one monitor but touching another could be reported as nearest to the adjacent one, which is clearly incorrect.

Bug: 1097049
Change-Id: I626c3ba07d5e084b1a9062b780b2f5d333e26ee0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2315663
Commit-Queue: Robert Liao <robliao@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793268}