Clear TestRootCerts in SSLUITestNoCert browser_test
Without this change, browser_tests will fail on ChromeOS with the
CertVerifierService enabled.

The tests that derive from SSLUITestNoCert rely on certain test root
certs not being present in the cert database.

In the tests, switches::kDisableTestCerts is added to the command line
which prevents EmbeddedTestServer::RegisterTestCerts() from being called
in the Network process.

However, when the CertVerifier is running in the browser process,
EmbeddedTestServer::RegisterTestCerts() is called when instantiating
an EmbeddedTestServer. So, for these two tests (only running on
Chromeos), just manually clear TestRootCerts in the browser process.

Bug: 1015134
Change-Id: I3416547a02657624be25a34ce91b26e912148fc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2327892
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Commit-Queue: Matthew Denton <mpdenton@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793295}