[WebLayer] Allow starting WebLayer shell in WebView compat mode
This adds an option to restart the shell in WebView compat mode, and
adds text in the version bar that says it's running in compat mode.

This should be very useful for testing things that happen only in
compat mode (like https://crbug.com/1111082).

Change-Id: I18fe9f2cf0df9a1789c40d650528579af0dc0a5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2330004
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793305}