Merge media::GpuVideoDecodeAcceleratorFactory::Create functions
There are a bunch of GL-related callbacks that these constructors
may take, and some GL-related callbacks may be added (or may not).
To that end, merge all of the callbacks into a
GpuVideoDecodeGLClient structure, and specify that.

Also start un-plumbing the AndroidOverlayMojoFactoryCB, since
it's not actually used.

R=sandersd
TBR=hiroh

Bug: 1108909
Change-Id: I5c47ef3992ca19930d462ac8e5d032667bd46c5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2320970
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793306}