cros: kShellWIndowId_ScreenRotation -> kShellWindowId_ScreenAnimation.
It was originally created for the screen rotation animation, but has
since been extended for use in desks animation and tablet mode
transition animation, and display highlight.

Display highlight is kind of an outlier as the new container name
doesn't quite match, but it matches the other three uses (and probably
future uses too).

Test: existing unittests
Bug: 1111445
Change-Id: I6bc25e39499f6bcd2d7dd8cdae4e878ad24a7298
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2329160
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793322}