inkdrop: Remove circle and round rect ink drop masks.
The last usage is FrameCaptionButton and it was a no-op because there
was a highlight path generator installed in the constructor.

PathInkDropMask is still used as the default for buttons that do not
have a mask override or a highlight path generator installed.

The base InkDropMask is still overridden by PathInkDropMask and the
browser app menu button's PulsingInkDropMask.

Test: existing tests
Bug: 1056490
Change-Id: I279d9c4edbe1258bdf414dfb04175f8d7fc4edd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2327957
Reviewed-by: Peter Boström <pbos@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793352}