Avoid late keyed services with TestingProfile
This patch migrates tests away from deprecated APIs and adopts
TestingProfile::Builder for tests in
/chrome/browser/ui/cocoa/history_menu_bridge_unittest.mm.

Rationale: creating or overriding keyed services after the profile has
been created is problematic and known to cause hard-to-debug test
flakiness, because it bypasses BrowserContextDependencyManager and often
leading to use-after-free.

This particular patch migrates the tests away from using a
HistoryService without an initialized database. This happened to be the
last test using TestingProfile that made use of this non-conventional
setup, which unblocks future simplifications.

However, changing this runs into flakiness issues because the history
backend tasks issue fatal errors if the underlying filesystem directory
is deleted too early. Therefore, BrowserWithTestWindowTest needs to be
updated to prevent the deletion of the temp dir before the ThreadPool
owned by TaskEnvironment is joined.

R=thakis@chromium.org

Bug: 1106699
Change-Id: Ie0db79270adb142395e9f12696705ad9029b28f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2316576
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793384}