device/fido: don't set attestation to kNone for Windows.
For device authenticators, the only thing that the attestation in the
request is used for is to select between enterprise attestation or not.
Thus kNone is a fine value for "not". However, the Windows authenticator
pays attention to this value and setting it to kNone caused f1c8b74669
to drop attestations on Windows.

BUG=1110242

Change-Id: I8b2f6d7aec56208bb9883122391724ddaa2ae796
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2330312
Commit-Queue: Adam Langley <agl@chromium.org>
Auto-Submit: Adam Langley <agl@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#793387}