[CryptAuth v2] Register with GCM and fetch metadata during service init
Register with GCM and fetch ClientAppMetadata as part of the DeviceSync
service initialization flow. Previously, GCM registration was performed
by the enrollment manager, and ClientAppMetadata was fetched on demand
by any class that needed it.

Note: Currently, the GCM registration ID is persisted across restarts,
and the ClientAppMetadata is lazily created and stored in memory by its
provider class; so there is already an assumption that the GCM
registration ID will never change and that the ClientAppMetadata will
not change during the user's session.

By fetching the ClientAppMetadata every time the service starts, we will
be able to detect if it has changes since the last successful
enrollment. For instance, a feature bit in the ClientAppMetadata might
be newly marked as "supported" behind a feature flag.

In a follow-up CL [3/4], we will directly pass the ClientAppMetadata to
other classes instead of passing the ClientAppMetadata provider. This
will remove a lot of technical complexity in place because of the
ClientAppMetadata provider's asynchronous API.

---

This is ClientAppMetadata change-detection CL [2/4]:

This chain of CLs relates to the end goal of re-enrolling if
ClientAppMetadata has changed since the last successful enrollment. For
example, when Phone Hub become supported, we will want to re-enroll.

We were previously only fetching ClientAppMetadata on demand, for
instance, right before a periodic enrollment. Now, we need to fetch the
data every time the service starts up to see if it changed. We continue
to make the assumption that the data will not change during a user
session.

Bug: 1105703
Change-Id: Id629244053b7cf557d01502d1a2ea32aa8029e74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2319328
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Josh Nohle <nohle@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793422}