Specify in the menu bar the flavor of Chrome
Right now, all flavors of Mac Chrome say “Chrome” in the menu bar.

Given that the idea of SxS is to run multiple Chrome instances at once,
make that text specify which flavor of Chrome is the active one.

The CFBundleName and CFBundleDisplayName in the strings files are
always the same as the value in the main Info.plist, so remove them.
Those override the value in the Info.plist, and the Info.plist
value is modified in packaging to have the “ Dev” and “ Beta”
suffixes.

Bug: 1111461
Change-Id: I793172d182b7bf6d680136cb357ef2cd627d3cb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2330075
Commit-Queue: Avi Drissman <avi@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793471}