media/gpu/vaapi_wrapper: refactor |implementation_type_|
I was just whipping up a CL to remove the unused |va_vendor_string|
param from IsBlockedDriver() when I noticed that there's quite a bit
of dance among classes to fill this value: The pre-sandbox singleton
VADisplay reads and keeps it as a string (but only valid between
InitializeOnce() and Deinitialize()); its client VASupportedProfiles
parses it into a VAImplementation enum and holds on to it, and further
classes use the enum value.

This CL just pushes the enum down to the VADisplay class, removing this
info from VASupportedProfiles entirely and simplifying the callsites.
Once parsed, it's kept for the lifetime of VADisplay.

No new functionality intended.

Bug: 1040291
Change-Id: I6f15cdc01977325faf37eada49404b8e0b7222da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2329736
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Auto-Submit: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793483}