Chrome OS - Minor clean up in LoginAuthUserView
LoginAuthUserView::SetAuthMethods has been misusing its AuthMethods
enum to control the visibility of the pin pad on the lock/login
screen.

Change the code so that it properly uses AuthMethods to indicate if
PIN authentication is possible and use an extra parameter to control
the visibility of the pin pad.

Bug: 1075994
Change-Id: I8f1a2eea009dc0d922f02bc9ced3f7005f773b83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2328860
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Commit-Queue: Renato Silva <rrsilva@google.com>
Cr-Commit-Position: refs/heads/master@{#793606}