[LayoutNG] Store the containing block offset
This change adds the containing block offset inside
NGPhysicalOutOfFlowPositionedNode and NGLogicalOutOfFlowPositionedNode.
This is part of the work for handling layout of positioned elements at
the fragmentainer level rather than at the containing block level.

A follow-up change will make use of this offset to compute where the
layout of an OOF positioned node should start in a multicolumn context.

Bug: 1079031
Change-Id: I2e3370a4ec0a4011203d0c3e916a64584aceaa2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2321405
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793618}