Finish the localdb move to PerformanceManager
This finishes the LocalDB migration to PerformanceManager. This
migration has been done by copying the code and now that the migration
is done it's also time to remove the previous version (it's hard to
keep both versions at the same time and to then remove the old version
in a later CL).

Most of these files are owned by chrisha@, except for the following
ones:

wfh:
  chrome/browser/ui/webui/discards/site_data.mojom
avi:
  chrome/browser/profiles/DEPS
  chrome/browser/profiles/chrome_browser_main_extra_parts_profiles.cc
  chrome/browser/usb/usb_tab_helper_unittest.cc
bcwhite:
  tools/metrics/histograms/histograms.xml

Bug: 961336, 1089370, 1049469, 977554
Change-Id: Id11ea6204d402204dd1f36a013e54bb64e7973b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2251747
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Chris Hamilton <chrisha@chromium.org>
Commit-Queue: Sébastien Marchand <sebmarchand@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793643}