[Paint Preview] Fix rect in PaintPreviewTracker
Link rects were previously recorded in a naive way that failed
to account for the overall transformation matrix that needs to be
applied to the rect. This CL rectifies that behavior by relying on
Annotations in the PaintOpBuffer and tracking transforms while walking
the buffer.

A similar fix for subframe rects might be needed (in a follow-up CL).

Bug: 1110227
Change-Id: I73cca7e4e398d724731f091c68910066d7b91785
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2327251
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Mehran Mahmoudi <mahmoudi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793650}