Create FlushForImageListener in CanvasResourceProvider
With deferred rendering it's possible for a drawImage operation on a
canvas to trigger a copy-on-write if another canvas has a read reference
to it. This can cause serious regressions due to extra allocations.
FlushForImageListener keeps a list of all active 2d contexts on a thread
and notifies them when one is attempting copy-on write. If the notified
context has a read reference to the canvas attempting a copy-on-write it
then flushes so as to make the copy-on-write unnecessary.

Results are promising!
https://aaronhk.users.x20web.corp.google.com/flush-for-image-performance.html?r=baseline&s=%25%CE%94avg&g=name&c=0

Bug: 1030108
Change-Id: Idc176053224634156b4775545dc56e8af54b0b1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2304251
Commit-Queue: Aaron Krajeski <aaronhk@chromium.org>
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793663}