Add banner on Chrome OS Settings for DeviceMinimumVersion policy
If an update is required as per DeviceMinimumVersion policy but the
device cannot be updated as it has reached end of life or auto
update expiration, the product requirement is to show a banner on the
top of Chrome OS Settings page.
This banner is to inform the user that device needs to be returned back
within the specified time frame. After this time frame has expired,
the user would not be able to use the device and hence, the close
button on the banner does not dismiss it forever but for just current
Settings page.
Screenshots:
https://screenshot.googleplex.com/PnKQdspE7G9
https://screenshot.googleplex.com/MGFnpRJSvrr
https://screenshot.googleplex.com/Ab2GT9bZcXH

Bug: 1110976
Change-Id: I69686b019eea524d162316d32ef483606e5154f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2314496
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Regan Hsu <hsuregan@chromium.org>
Commit-Queue: Saurabh Nijhara <snijhara@google.com>
Cr-Commit-Position: refs/heads/master@{#793667}